backend/requirements.txt
# SSL Monitor Backend Dependencies

# ============================================
# Core Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# Database
# ============================================
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ============================================
# Security & Authentication
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.1

# ============================================
# Validation & Serialization
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# Rate Limiting & Security
# ============================================
slowapi==0.1.9

# ============================================
# SSL Certificate Handling
# ============================================
cryptography==41.0.7
certifi==2023.7.22
requests==2.31.0

# ============================================
# Logging & Monitoring
# ============================================
python-json-logger==2.0.7

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
setuptools==68.2.2

# ============================================
# Development Dependencies (optional)
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.1
# flake8==6.1.0
# mypy==1.7.1